# doc/api.yaml
openapi: 3.0.0
info: 
  description: WASAText exam project Rest API
  title: WASAText
  version: 1.0.0
servers:
- url: http://localhost:3000
components:
  schemas:
    user:
      title: user
      description: "This object represent the unique name of the user (\"user\")."
      type: object
      properties:
        u_id:
          description: unique identifier
          type: integer
        u_username:
          description: Unique name of the user
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 32
        u_profileImage:
          description: profile picture of the user
          type: string
          pattern: '^.*?$'
          format: binary
          minLength: 1
          maxLength: 1024
        u_bio:
          description: biography of the user
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 256
        u_firstName:
          description: the real name of the user 
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 32
        u_lastName:
          description: the real surname of the user
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 32
...
  securitySchemes:
    BearerAuth: 
      type: http 
      scheme: bearer 
      bearerFormat: JWT   
...
paths:
...
  /users/{id}/username:
    post:
      tags: ["user profile"]
      summary: set the username of the user
      description: |
         sets the username
      operationId: setMyUserName 
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      requestBody: 
        description: new username
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                un_name:
                  description: temp
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
        required: true
      responses:
        '201': 
          description: username successfully set 
          content:
            application/json:
              schema:
                description: temp
                type: object
                properties:
                  un_username:
                    description: temp
                    type: string
                    pattern: '^.*?$'
                    minLength: 1
                    maxLength: 32
        '401':
          description: not logged in
        '403':
          description: username already exist
        '406': 
          description: invalid username
...
  /users/{id}/conversations/{conversationId}/messages/{messageId}/comments:
    parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: conversationId
        in: path
        description: id conversation
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: messageId
        in: path
        description: id message
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
    post:
      tags: ["message management"]
      summary: comment another message
      description: comment another message 
      operationId: commentMessage
      security:
        - BearerAuth: []   
      requestBody: 
        description: message to comment
        content:
          multipart/form-data:
            schema:
              description: temp
              type: object
              properties:
                messageText:
                  description: the text part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
                photoText:
                  description: the photo part of the message
                  type: string
                  pattern: '^.*?$'
                  minLength: 1
                  maxLength: 32
        required: true
      responses:
        '204': 
          description: message commented
        '401':
          description: not logged in
        '406': 
          description: invalid text
    delete:
      tags: ["message management"]
      summary: uncomment another message
      description: uncomment another message
      operationId: uncommentMessage
      security:
        - BearerAuth: []   
      responses:
        '204': 
          description: comment successfully uncommented
        '401':
          description: not logged in
        '403':
          description: unauthorized to delete this message
        '404':
          description: The requested resource could not be found 
...
  /users/{id}/photo:
    post:
      tags: ["user profile"]
      summary: set the photo of the user
      description: change the photo of the user, if successful returns the user
      operationId: setMyPhoto
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      requestBody:
        description: the new photo
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                smp_photo:
                  description: temp
                  type: string
                  format: binary
                  minLength: 1
                  maxLength: 1024
        required: true
      responses:
        '201': 
          description: photo changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: not logged in
  /users/{id}/conversations/{conversationId}/groupphoto:
    post:
      tags: ["chat management"]
      summary: change the photo of the group
      description: change the photo of the group, if successful returns the group
      operationId: setGroupPhoto
      security:
        - BearerAuth: []   
      parameters:
      - name: id
        in: path
        description: id utente
        required: true
        schema:
          description: temp
          type: integer
          minimum: 0
      - name: conversationId
        in: path
        description: id conversation
        required: true
        schema:
          description: temp
          type: object
          properties:
            sgp_con_id:
              description: temp
              type: integer
      requestBody:
        description: the new photo and the group
        content:
          application/json:
            schema:
              description: temp
              type: object
              properties:
                sgp_photo:
                  description: temp
                  type: string
                  format: binary
                  minLength: 1
                  maxLength: 1024
        required: true
      responses:
        '204': 
          description: photo changed
        '401':
          description: not logged in
        '403':
          description: not authorized